+++ CHANGELOG.md
@@ -40,7 +40,9 @@ Il formato utilizzato è basato sulle linee guida di [Keep a Changelog](http://k
 ### Aggiunto (Added)
  - Aggiunta nel calendario della Dashboard visualizzazione dei preventivi pianificabili in corrispondenza alla data di accettazione e conclusione.
  - Aggiunta impostazione per la visualizzazione delle ore nella stampa intervento (Decimale, Sessantesimi).
-
+ - Aggiunta possibilità di selezionare la sede di partenza della merce in fase di aggiunta articolo da un'attività
+ - Sostituito plugin **Componenti** nel modulo Impianti con la possibilità di inserire gli articoli di magazzino
+ 
 ### Fixed
  -